move proximity out of contrib and rename proximity_forest
