+++ sktime/contrib/experiments.py
@@ -17,7 +17,7 @@ import sktime.classifiers.dictionary_based.boss as db
 import sktime.classifiers.frequency_based.rise as fb
 import sktime.classifiers.interval_based.tsf as ib
 import sktime.classifiers.distance_based.elastic_ensemble as dist
-import sktime.contrib.distance_based.proximity as pf
+import sktime.classifiers.distance_based.proximity_forest as pf
 from sktime.utils.load_data import load_from_tsfile_to_dataframe as load_ts
 
 __author__ = "Anthony Bagnall"
